requirements.txt
# --- 核心深度学习框架 ---
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
transformers==4.33.2
timm==0.4.12
accelerate==0.34.2
peft==0.12.0
fairscale==0.4.4
einops==0.8.0

# --- 图像处理与退化相关 ---
opencv-python-headless==4.5.5.64
Pillow==10.4.0
scikit-image==0.21.0
scipy==1.10.1
basicsr==1.4.2

# --- 数据处理与存储 ---
numpy==1.24.4
pandas==2.0.3
h5py==3.11.0
webdataset==0.2.100
decord==0.6.0
iopath==0.1.10
addict==2.4.0
omegaconf==2.3.0
ml_collections==0.1.1

# --- 评测与 NLP 工具 ---
pycocoevalcap==1.2
pycocotools==2.0.7
nltk==3.9.1
spacy==3.7.6
sentencepiece==0.2.0
regex==2024.9.11

# --- 实用工具 ---
tqdm==4.66.5
PyYAML==6.0.2
requests==2.32.3
huggingface-hub==0.24.7